Initial implementation of reversed()
The test for it doesn't work as-is due to an issue of r.append() adding None
to the data stack, which screws up the for-loop logic. We have determined
that the basic test works without this, but we're going to freeze this
amount of the implementation while chasing down that problem.

Created a common iterator generator to help create iterators using composition.