Single, isolated commands no longer put things on the data stack
This behavior caused a problem with for-loops. The polluted item would screw
up the loop commands. It would think the result of the function was the
iterator it was supposed to use on the second pass through the loop.

This change has broken the REPL. We were using these elements to determine
what to print. CPython checks if we're in a REPL and issues a PRINT_EXPR
opcode to these situations. The interpreter now looks for a print built-in
and does nothing if it doesn't have one. The REPL for testing has one.
Users would want to add their own if they want to capture this output.

InterpretAsync still returns a string for preparation and compilation errors.
However, it's assumed that they will add a built-in for print that they'll use
for PRINT_EXPR.

Added the most cursory of documentation about the print function to manual.