Added a list() constructor
This was needed as part of supporting heapq. The list() function can either take
nothing and produce an empty list, or take an iterable and make a list of out it.
We support both forms now.

While at it, we added tests for a dict() constructor that we will want to implement
some day.