Imported functions now retain module globals through PyFunction globals
The problem was that a module with a function calling another function could
not resolve the call. This is because the interpreter was carrying globals
down from the original root module instead of switching globals to the new
module's globals. There was nothing in place to even track the module's
globals separately, nor a place to give this information to the code being
called.

CPython has PyFunction which encapsulates all these code objects we run.
This object has a globals property where the module-level globals would be
given. We created our own PyFunction with a global property. CodeObjects
are contained inside them and many CodeObject parameters became PyFunctions
internally.

We then had to do a little extra effort with our frame preparation to ensure
module globals and locals are tied together when they are created, and that
the built PyFunctions inside the module all have access to the module's
globals when the functions are being made.

The string module loader was enhanced from the initial drafts to properly
handle the changes. The file-based loader also needed some revisions.