Added two tests to demonstrate short circuiting doesn't happen (bug)
Heapq tests showed that logic short circuiting doesn't happen. So if you do
something typical like:

```
if len(a) > b and a[b] > a[b-1]:
```
or similar, then it'll evaluate both statements even if the first is false.
This causes problems with these statements that are otherwise legal. Or
statements always need to short circuit (and don't).