Initial replacement of and logic with short-circuit jump tests
This fixes issues where we try to run all statements in the and block even if
the first test is false. Put another way, this implements a basic short-circuit
and. A sensitive issue here is how well this works with more than just a simple
a AND b kind of test; we need to verify with multiple fields.

Short-circuit OR is also not implemented yet, but will be a straight-forward
next step.