A superficial attempt at explaining how to get a functional build
This should address issue #4. However, I think this will ultimately be improved
when I start to run commits through an external QA harness. That's because that
harness will have to fight this every time.

One solution may be to just commit the generated files too, which is kind of
yuck.