Actually did proper merging with temporary list
This whole swapping thing I was doing was just wrong. I looked up merge sort
and realized I needed to work from a temporary space. There's a version that
doesn't need that space but requires I shift all the values right and insert
as I go, which is its own can of worms.