Officially breaking the Call() API to handle defaultoverrides
Throwing in **kwargs while I'm at it.

I had to take the plunge because I discovered I make Call()'s all over the
place that need to also resolve, and I'd rather get that under control now and
for all than to put a resolve call in front of each call I have.