FindBestMethodMatch reports what generics it stripped now
The injector no longer thinks about generics. It doesn't sweat it at all. This
seems to thread the needle I am having. I still have 10 failing tests but
but whack-a-mole problem is severely reduced... so is the amount of code I
have to peddle around.

FindBestMethodMatch returns FoundBestMethod which includes what generics
were stripped if the method had to be monomorphed.