Got tracebacks working for embedded PyExceptions
PyExceptions created inside .NET code weren't getting any tracebacks. The
solution was to create a helper for manipulating the traceback. It started from
some existing code but it had to be enhanced for cases where the traceback
was null, and for cases where the traceback already had the current line and
code at the top (happens with overlap of checks).

This gave us more exposure into the problem in quicksort but wasn't a silver
bullet.