Added a first-pass name scope resolution parser
This stemmed from having to support nonlocal along with solving some issues with
the smallest_change test. After various discussions, I figured out I had an issue
with dealing with enclosing variables when I had a single-pass parser. If I had to
promote a local to some kind of enclosed variable (cell variable or whatever) later,
I would have to effectively rewrite a bunch of code on-the-fly. What I was told is
that CPython itself does this in two passes, so I did that two. So this adds a
"Variable Scan Visitor" that creates a tree of names and their scopes. It figures
out the scope of all used names at each level.

All of this worked culminated in some more work to properly support cell variables,
which dragged in support for closures. Some issues in other integration tests also
went away while doing this. I had to ask a bunch of questions about variable
resolution in Python and needed multiple attempts to get different parts of it
right.

# Conflicts:
#	CloacaCodeTests/fixing.md
#	CloacaInterpreter/Visitors/VariableScanVisitor.cs
#	CloacaTests/VariableScanVisitorTests.cs