Unintegrated draft of a variable scope visitor
I learned that CPython uses a 2-part parsing process where the first part determines
the scope of named items. This is how it determines nonlocal stuff in particular.
This visitor isn't actually used in the main code yet, but has been tested against
some scenarios and should be ready to try.

Integrating this will be essential for properly handling enclosed scope and emitting
DEREF instructions.

I'm not 100% convinced it's completely correct yet, but I expect to find out details
when I start integrating.