Started plumbing variable visitor errors out
I think it's gross. It returns a type that becomes an exception anyways. I
thought ANTLR4 had some method of stopping stuff without throwing an exception,
but nope! So I may refactor this a bit before moving on. However, it does err
on nonlocals trying to escape a class now.