Finally passing ClassWithGLC
This was one of the "hard ones" and I need to review what I did. The
implementation adding class and method names as variables to scope out, and
I need to double check that is correct. If so, I need to keep this test alive
while trying to enable "simpler" tests.