Now using NAME resolution for unbound locals with reads
We were injecting __name__ as a module level variable and the normal convention
is to just let LEGB resolution take its natural course. Instead, we were
throwing an error if we couldn't resolve the variable. Not good.

Unfortunately, this passed at least one more test but not as many as I would
have hoped.