Free variables now are marked in between layers that don't reference them
This is needed when the free variable is used multiple layers down in multiple
function definitions. We carry it as a free variable.

The two closure tests STILL don't pass but we now should be giving everything
enough information to actually succeed so the debugging should be less
puzzling (hehehe).