Got basic closures to work (at expense of a few recursive tests)
This was a whole pile of changes that all together enabled closures. I added
some helpers to make it clearer when the closure was getting involved.
There had to be separate of logic between cell and free variables.

I'll have to figure out what I broke in the recursive examples next.

Not sure if LOAD/STORE_DEREF should always look at freenames