Added Schedule command that prepare variables before scheduling
I was setting the variables *after* scheduling, and doing that all over the
place. So I encapsulated in the scheduler. The code is still gross and I wish
to do it better. This is because if how strange it looks to test for root
context (globals == local) and use SetIfVariableExists(). At the least, I need
a short essay in that section to explain why it's done that way if that's just
how it's done.